+++ client/src/features/advancedSearch/Group.tsx
@@ -10,7 +10,7 @@ import { addFieldSelection } from '../../redux/slices/advancedSearchSlice'
 import AlignedDiv from '../../styles/features/advancedSearch/AlignedDiv'
 import CollapseButton from '../../styles/shared/CollapseButton'
 import VerticalLine from '../../styles/features/advancedSearch/VerticalLine'
-import { pushSiteImproveEvent } from '../../lib/siteImprove'
+import { pushClientEvent } from '../../lib/pushClientEvent'
 
 import AddButton from './AddButton'
 import CollapseContainer from './CollapseContainer'
@@ -92,7 +92,7 @@ const Group: React.FC<IGroup> = ({
         <VerticalLine className="groupVerticalLine">
           <CollapseButton
             onClick={() => {
-              pushSiteImproveEvent(
+              pushClientEvent(
                 'Collapse Button',
                 open ? 'Open' : 'Close',
                 'Advanced Search',