+++ client/src/features/search/AdvancedSearchButton.tsx
@@ -13,7 +13,7 @@ import { searchScope } from '../../config/searchTypes'
 import { getParamPrefix } from '../../lib/util/params'
 import Tooltip from '../common/Tooltip'
 import EntityResultsDescription from '../cms/EntityResultsDescription'
-import { pushSiteImproveEvent } from '../../lib/siteImprove'
+import { pushClientEvent } from '../../lib/pushClientEvent'
 
 interface IProps {
   setIsError: (x: boolean) => void
@@ -29,7 +29,7 @@ const AdvancedSearchButton: React.FC<IProps> = ({ setIsError, id }) => {
 
   const handleSelect = (entityType: string): void => {
     const searchString = value || ''
-    pushSiteImproveEvent(
+    pushClientEvent(
       'Search Switch',
       'Selected',
       `To Advanced Search ${entityType}`,