+++ client/src/features/search/SearchBox.tsx
@@ -13,7 +13,7 @@ import {
 } from '../../redux/slices/simpleSearchSlice'
 import theme from '../../styles/theme'
 import LoadingSpinner from '../common/LoadingSpinner'
-import { pushSiteImproveEvent } from '../../lib/siteImprove'
+import { pushClientEvent } from '../../lib/pushClientEvent'
 
 const StyledSearchBox = styled.div`
   display: flex;
@@ -116,7 +116,7 @@ const SearchBox: React.FC<{
           inputRef.current!.value = ''
           setIsError(false)
           setIsLoading(false)
-          pushSiteImproveEvent('Search Button', 'Submit', 'Simple Search')
+          pushClientEvent('Search Button', 'Submit', 'Simple Search')
           navigate(
             {
               pathname: `/view/results/${tab}`,