+++ client/src/test/integration/PersonPage.spec.tsx
@@ -3,14 +3,14 @@ import React from 'react'
 
 import AppRender from './utils/AppRender'
 import personMockApi from './utils/personMockApi'
-import siteImproveMock from './utils/mockSiteImprove'
+import eventTrackingMock from './utils/eventTrackingMock'
 
 describe('Person page', () => {
   const page = '/view/person/mock-person'
 
   beforeEach(async () => {
     personMockApi()
-    siteImproveMock()
+    eventTrackingMock()
   })
 
   describe('About', () => {