+++ client/CHANGELOG.md
@@ -2,7 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.28.0 - Unreleased
+## v1.29.0 - Unreleased
+
+### Added
+
+- Added Objects and Works Influenced Tabs to Person & Group pages ([#87](https://github.com/project-lux/lux-frontend/issues/87)).
+- Added results scope to the refine label ([#134](https://github.com/project-lux/lux-frontend/issues/134)).
+- Added component term facets to concept results ([#68](https://github.com/project-lux/lux-frontend/issues/68)).
+
+### Changed
+
+### Removed
+
+### Fixed
+
+## v1.28.0 - 2024-6-10
 
 ### Added
 
@@ -145,7 +159,9 @@ All notable changes to this project will be documented in this file.
 - Added People Active Here accordion to place pages ([#34](https://github.com/project-lux/lux-frontend/issues/34)).
 - Added Place with Type on concept pages ([#107](https://github.com/project-lux/lux-frontend/issues/107)).
 - Added Events with Type on concept pages ([#128](https://github.com/project-lux/lux-frontend/issues/128)).
-- Added entity icon to About/Depicts ([#86](https://github.com/project-lux/lux-frontend/issues/86)).
+- Added entity icon to About/Depicts ([#84](https://github.com/project-lux/lux-frontend/issues/84)).
+- Added index.html.template and updated docs ([#77](https://github.com/project-lux/lux-frontend/issues/77)).
+- Added event tracking abstraction ([#78](https://github.com/project-lux/lux-frontend/issues/78)).
 
 ### Changed